+++ src/utils/convert.rs
@@ -1,0 +1,5 @@
+pub fn convert(value: String, to_bin: bool, to_hex: bool, to_dec: bool) {
+
+    println!( "{}", value);
+
+}