+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - Memory funcion was added, with 3 flags.
 - Added frist README.
 - Added the struct for convert.
+- Minor fixes in how prints information